Lobby, Game screen created
Created searchable lobby screen
Game screen with canvas, chat and score

+++ server/src/main/java/com/kruna1pate1/pictionaryserver/model/DrawData.java
@@ -21,5 +21,5 @@ public class DrawData {
 
     private CanvasAction action;
 
-    private Float color;
+    private Integer color;
 }
